% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_assign}
\alias{h9_assign}
\title{Assign}
\usage{
h9_assign(h, column = NULL, array = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{column}{Name of the new column}

\item{array}{The array to be added to the table}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Attach a new column
}
