% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_bar_chart}
\alias{h9_bar_chart}
\title{Bar}
\usage{
h9_bar_chart(
  h,
  x = NULL,
  y = NULL,
  type = NULL,
  orientation = NULL,
  palette = NULL,
  fontsize = NULL,
  tickrotation = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The variable that horizontal axis}

\item{y}{The variable that should be on the vertical axis.}

\item{type}{One of 'grouped' or 'stacked'. In grouped mode, bars are placed next to each other, in stacked mode bars are placed above each other.}

\item{orientation}{}

\item{palette}{Colors to use for the different levels of the y variable. Should be one of the valid d3.js color palettes.}

\item{fontsize}{Size of the font to be used in the x and y axes.}

\item{tickrotation}{The angle at which to place the x-axis labels}

\item{marginleft}{the left margin}

\item{marginbottom}{the bottom margin}
}
\value{
A list with the pipeline specification.
}
\description{
Our first chart type needs no introduction
}
