% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_visualizations.R
\name{h9_bubble_chart}
\alias{h9_bubble_chart}
\title{Bubbles}
\usage{
h9_bubble_chart(h = NULL, label = NULL, size = NULL, palette = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{label}{The column in the dataframe who's values should be used as labels on the circles}

\item{size}{The column in the dataframe who's values should the area of each circle be propotional to.}

\item{palette}{the D3 Palette to determine the color scheme to use}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Bigger values -> bigger bubbles
}
