% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_checkbox}
\alias{h9_checkbox}
\title{Checkbox}
\usage{
h9_checkbox(uid, label = NULL, checked = NULL, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{label}{Specifies the label for this checkbox}

\item{checked}{Specifies the checked for this checkbox}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a checkbox control
}
