% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h9_city_to_map}
\alias{h9_city_to_map}
\title{Coordinates from City}
\usage{
h9_city_to_map(h, city = NULL, state = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{city}{}

\item{state}{}
}
\value{
A list with the pipeline specification.
}
\description{
Create latitude and longitude columns from a pair of US city and state columns
}
