% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_convert}
\alias{h9_convert}
\title{Convert}
\usage{
h9_convert(
  h,
  field = NULL,
  dataType = NULL,
  timeConverter = NULL,
  charactersToRemove = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{field}{The name of the column to convert}

\item{dataType}{The target data type}

\item{timeConverter}{an optional parameter to help convert date-times}

\item{charactersToRemove}{a string of characters to remove}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Apply an expression to the values in a specific column
}
