% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_derive}
\alias{h9_derive}
\title{Derive}
\usage{
h9_derive(h, column = NULL, expression = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{column}{The name of the new column. If this is a column that already exists in the dataframe, the derived column replaces the exisiting one.}

\item{expression}{The expression based on which the new columns are derived}
}
\value{
A list with the pipeline specification.
}
\description{
Create a new column or replace an existing one via an expression
}
