% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_dot_plot}
\alias{h9_dot_plot}
\title{Dot Plot}
\usage{
h9_dot_plot(
  h,
  x = NULL,
  color = NULL,
  palette = NULL,
  dotsize = NULL,
  ticks = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{color}{}

\item{palette}{}

\item{dotsize}{}

\item{ticks}{}

\item{fontsize}{}

\item{marginleft}{}

\item{marginbottom}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Stack data into columns of dots
}
