% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_dropdown}
\alias{h9_dropdown}
\title{Dropdown}
\usage{
h9_dropdown(uid, values = NULL, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{values}{Specifies the values for this dropdown}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a dropdown control
}
