% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_error_bar}
\alias{h9_error_bar}
\title{Candles}
\usage{
h9_error_bar(
  h,
  x = NULL,
  min = NULL,
  max = NULL,
  open = NULL,
  close = NULL,
  levels = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{marginbottom}{}
}
\value{
A list with the pipeline specification.
}
\description{
Explain financial trends with a candlestick chart
}
