% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_error_chart}
\alias{h9_error_chart}
\title{Candles}
\usage{
h9_error_chart(
  h,
  x = NULL,
  min = NULL,
  max = NULL,
  open = NULL,
  close = NULL,
  levels = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{the column in the dataframe that is the contains the values for the x coordinates}

\item{min}{the column in the dataframe that is the contains the values for the minimum at each x coordinate}

\item{max}{the column in the dataframe that is the contains the values for the maximum at each x coordinate}

\item{open}{the column in the dataframe the contains the values at opening at each x coordinate}

\item{close}{the column in the dataframe the contains the values at close m at each x coordinate}

\item{levels}{Additional step parameter.}

\item{fontsize}{the font size}

\item{marginleft}{The left margin}

\item{marginbottom}{the bottom margin}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Explain financial trends with a candlestick chart
}
