% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{h9_estimate_pose}
\alias{h9_estimate_pose}
\title{Pose}
\usage{
h9_estimate_pose(h, images = NULL, model = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{images}{Additional step parameter.}

\item{model}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Apply automated post-process motion capture dots to an image
}
