% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{h9_export_text}
\alias{h9_export_text}
\title{Text}
\usage{
h9_export_text(h, field = NULL, separator = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{field}{}

\item{separator}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Export a dataset into a delimited string
}
