% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_facets}
\alias{h9_facets}
\title{Facets}
\usage{
h9_facets(
  h,
  x = NULL,
  y = NULL,
  facets = NULL,
  color = NULL,
  chartType = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{facets}{}

\item{color}{}

\item{chartType}{}

\item{palette}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
A chart of charts organized in a grid
}
