% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_fetch}
\alias{h9_fetch}
\title{Fetch}
\usage{
h9_fetch(h, url = NULL, resize = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{}

\item{resize}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Create new data columns by processing a URL column with a Hal9 server worker
}
