% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_file}
\alias{h9_file}
\title{File Input}
\usage{
h9_file(uid, dataType = NULL, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{dataType}{Specifies the dataType for this fileinput}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a file input control
}
