% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_filter}
\alias{h9_filter}
\title{Filter}
\usage{
h9_filter(h, field = NULL, expression = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{field}{The column on which to filter}

\item{expression}{the criteria on which to filter the rows}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Keep only the rows that satisfy a given expression for a specific column
}
