% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_funnel}
\alias{h9_funnel}
\title{Funnel}
\usage{
h9_funnel(
  h,
  stage = NULL,
  value = NULL,
  label = NULL,
  fontSize = NULL,
  showPercentSelection = NULL,
  funnelType = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{stage}{}

\item{value}{}

\item{label}{}

\item{fontSize}{}

\item{showPercentSelection}{}

\item{funnelType}{}

\item{palette}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Communicate how data narrows down
}
