% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_funnel_chart}
\alias{h9_funnel_chart}
\title{Funnel}
\usage{
h9_funnel_chart(
  h,
  stage = NULL,
  value = NULL,
  label = NULL,
  fontSize = NULL,
  showPercentSelection = NULL,
  funnelType = NULL,
  palette = NULL,
  rebind = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{stage}{The column in the dataframe that contains all the stages of the process}

\item{value}{The column in the dataframe that contains the values corresponding to each stage of the process}

\item{label}{The column in the dataframe that should be used to create labels of the stages}

\item{fontSize}{The font size}

\item{showPercentSelection}{Boolean on whether to show the percentage in each step}

\item{funnelType}{The type of the funnel to make. Options are 2d, 3d and flat.}

\item{palette}{Colors to use for the different levels of the y variable. Should be one of the valid d3.js color palettes.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Communicate how data narrows down
}
