% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_heatmap}
\alias{h9_heatmap}
\title{Heatmap}
\usage{
h9_heatmap(
  h,
  x = NULL,
  y = NULL,
  value = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{value}{}

\item{palette}{}

\item{fontsize}{}

\item{marginleft}{}

\item{marginbottom}{}
}
\value{
A list with the pipeline specification.
}
\description{
Draw attention to larger values with more vibrant colors in this 2D map
}
