% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_heatmap_chart}
\alias{h9_heatmap_chart}
\title{Heatmap}
\usage{
h9_heatmap_chart(
  h,
  x = NULL,
  y = NULL,
  value = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The column in the dataframe that defines the x coordinates of the marks}

\item{y}{The column in the dataframe that defines the y coordinates of the marks}

\item{value}{The column in the dataframe that defines the intensity of the colors of the marks}

\item{palette}{Additional step parameter.}

\item{fontsize}{The font size}

\item{marginleft}{The left margin}

\item{marginbottom}{The bottom margin}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Draw attention to larger values with more vibrant colors in this 2D map
}
