% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_histogram_chart}
\alias{h9_histogram_chart}
\title{Histogram}
\usage{
h9_histogram_chart(
  h,
  x = NULL,
  histfunc = NULL,
  histnorm = NULL,
  barmode = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The name of the column in the input dataframe that who's distribution to be visualized}

\item{histfunc}{The function used to aggregate the values collected in each bin for summarization.}

\item{histnorm}{The aggregation method to apply on outputs of the aggregation functions.}

\item{barmode}{One of stacked or overlaid, which controls the manner in which multiple distributions selected in x are visualized.}

\item{palette}{The D3 palette used to control the colors of each of the distributions in x.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Show further granularity within buckets via optional color palettes
}
