% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{h9_import_airbnb}
\alias{h9_import_airbnb}
\title{Airbnb}
\usage{
h9_import_airbnb(h, url = NULL, maxReviews = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{}

\item{maxReviews}{}
}
\value{
A list with the pipeline specification.
}
\description{
Gather comments from an Airbnb listing by using a Hal9 server worker
}
