% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{h9_import_bigquery}
\alias{h9_import_bigquery}
\title{BigQuery}
\usage{
h9_import_bigquery(
  h,
  clientId = NULL,
  projectNumber = NULL,
  query = NULL,
  dbLocation = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{clientId}{Additional step parameter.}

\item{projectNumber}{Additional step parameter.}

\item{query}{Additional step parameter.}

\item{dbLocation}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Query data from a Google Cloud BigQuery data warehouse
}
