% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_import_graphql}
\alias{h9_import_graphql}
\title{GraphQL}
\usage{
h9_import_graphql(h, url = NULL, query = NULL, extract = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{Additional step parameter.}

\item{query}{Additional step parameter.}

\item{extract}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Query data from a GraphQL endpoint
}
