% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_import_mysql}
\alias{h9_import_mysql}
\title{MySQL}
\usage{
h9_import_mysql(
  h,
  host = NULL,
  user = NULL,
  password = NULL,
  database = NULL,
  query = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{host}{Additional step parameter.}

\item{user}{Additional step parameter.}

\item{password}{Additional step parameter.}

\item{database}{Additional step parameter.}

\item{query}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Query data from a MySQL database by using a Hal9 server worker
}
