% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{h9_import_sheet}
\alias{h9_import_sheet}
\title{Google Sheets}
\usage{
h9_import_sheet(
  h,
  url = NULL,
  documentAccess = NULL,
  sheet = NULL,
  clientId = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{}

\item{documentAccess}{}

\item{sheet}{}

\item{clientId}{}
}
\value{
A list with the pipeline specification.
}
\description{
Import data from a Google Sheets spreadsheet
}
