% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{h9_import_twitter}
\alias{h9_import_twitter}
\title{Twitter}
\usage{
h9_import_twitter(
  h,
  search = NULL,
  key = NULL,
  secret = NULL,
  type = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{search}{}

\item{key}{}

\item{secret}{}

\item{type}{}
}
\value{
A list with the pipeline specification.
}
\description{
Search Twitter with a Hal9 server worker
}
