% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_import_video}
\alias{h9_import_video}
\title{Video Frames}
\usage{
h9_import_video(h, file = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{file}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Grab frames from a video file
}
