% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_impute}
\alias{h9_impute}
\title{Impute}
\usage{
h9_impute(h, field = NULL, method = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{field}{List of columns in which to replace missing values}

\item{method}{the function to use to impute the missing values. Currently available options are 'max', 'min', 'mean' median', 'zero'. Default is 'zero'}
}
\value{
A list with the pipeline specification.
}
\description{
Fill in a column's missing data with the value of an aggregation
}
