% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_map}
\alias{h9_map}
\title{Map}
\usage{
h9_map(h, lon = NULL, lat = NULL, size = NULL, label = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{lon}{}

\item{lat}{}

\item{size}{}

\item{label}{}
}
\value{
A list with the pipeline specification.
}
\description{
Clearly map out Earthly datasets that have columns for latitude and longitude
}
