% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_markdown}
\alias{h9_markdown}
\title{Markdown}
\usage{
h9_markdown(uid, markdown = NULL, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{markdown}{Specifies the markdown for this md}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Enables to render arbitrary markdown
}
