% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_message}
\alias{h9_message}
\title{Message}
\usage{
h9_message(
  uid,
  message = NULL,
  title = NULL,
  mtype = NULL,
  showIcon = NULL,
  on_update = NULL,
  ...
)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{message}{Specifies the message for this message}

\item{title}{Specifies the title for this message}

\item{mtype}{Specifies the mtype for this message}

\item{showIcon}{Specifies the showIcon for this message}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a message element
}
