% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_network}
\alias{h9_network}
\title{Network}
\usage{
h9_network(h, from = NULL, to = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{from}{}

\item{to}{}
}
\value{
A list with the pipeline specification.
}
\description{
Dazzle your colleagues with this colorful animated network graph
}
