% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_visualizations.R
\name{h9_network_chart}
\alias{h9_network_chart}
\title{Network}
\usage{
h9_network_chart(h = NULL, from = NULL, to = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{from}{Column containing the origin vertices of all the edges in the graph}

\item{to}{Column containing the target vertices of all the edges in the graph}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Dazzle your colleagues with this colorful animated network graph
}
