% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_number}
\alias{h9_number}
\title{Number Input}
\usage{
h9_number(uid, label = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{...}{Other h9 parameters.}
}
\description{
Embed a number input control
}
