% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_pivot}
\alias{h9_pivot}
\title{Pivot}
\usage{
h9_pivot(
  h,
  rows = NULL,
  columns = NULL,
  values = NULL,
  summarizer = NULL,
  rebind = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{rows}{The column whose unique values should serve as the rows of the new dataframe}

\item{columns}{The column whose unique values should serve as the columns of the new dataframe}

\item{values}{The columns whose values should be collected to serve as the individual cells of the table}

\item{summarizer}{the method to aggregate the collected values}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Pivot columns into a cross-tabulation
}
