% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_plotly}
\alias{h9_plotly}
\title{Plotly}
\usage{
h9_plotly(
  h,
  x = NULL,
  y = NULL,
  chartType = NULL,
  dataSizes = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{chartType}{}

\item{dataSizes}{}

\item{palette}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Choose from a variety of Plotly visualizations
}
