% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{h9_predict_image}
\alias{h9_predict_image}
\title{Image Classification}
\usage{
h9_predict_image(h, url = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Use MobileNet's Image Classification to attempt to identify what's in an image
}
