% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{h9_predict_lstm}
\alias{h9_predict_lstm}
\title{LSTM}
\usage{
h9_predict_lstm(
  h,
  prediction = NULL,
  window = NULL,
  units = NULL,
  epochs = NULL,
  predictions = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{prediction}{}

\item{window}{}

\item{units}{}

\item{epochs}{}

\item{predictions}{}
}
\value{
A list with the pipeline specification.
}
\description{
Predict using a long-short-term-memory model
}
