% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{h9_predict_sentiment}
\alias{h9_predict_sentiment}
\title{Sentiment}
\usage{
h9_predict_sentiment(h, sentiment = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{sentiment}{}
}
\value{
A list with the pipeline specification.
}
\description{
Ask a neural net to predict how positively the text in a dataset will be perceived
}
