% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_radial_chart}
\alias{h9_radial_chart}
\title{Radial}
\usage{
h9_radial_chart(
  h,
  x = NULL,
  y = NULL,
  wafflesizelabel = NULL,
  palette = NULL,
  rebind = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The column containing the labels of the charts}

\item{y}{the column containing the values the areas they occupy in the rectagular area should be propotional to}

\item{wafflesizelabel}{The size of the large rectangle}

\item{palette}{the D3 Palette to determine the color scheme to use}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
A circular bar chart
}
