% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_rawhtml}
\alias{h9_rawhtml}
\title{HTML}
\usage{
h9_rawhtml(uid, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Enables to render arbitrary HTML
}
