% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_read_csv}
\alias{h9_read_csv}
\title{CSV}
\usage{
h9_read_csv(h, file = NULL, separator = NULL, skip = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{file}{Either the URL for the csv or a local file.}

\item{separator}{A single-character delimiter string between column values (default ',')}

\item{skip}{The number of lines to skip (default 0) before reading data}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Import a CSV formatted dataset
}
