% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_read_excel}
\alias{h9_read_excel}
\title{Excel}
\usage{
h9_read_excel(h, file = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{file}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Import a dataset from a sheet of an Excel file
}
