% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_read_sqlite}
\alias{h9_read_sqlite}
\title{SQLite}
\usage{
h9_read_sqlite(h, file = NULL, query = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{file}{}

\item{query}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Query data from a SQLite database
}
