% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_roll_sum}
\alias{h9_roll_sum}
\title{Rolling Sum}
\usage{
h9_roll_sum(h, column = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{column}{The column for which to calculate the rolling sum.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Create new columns from the running totals of others
}
