% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transforms.R
\name{h9_sample}
\alias{h9_sample}
\title{Sample}
\usage{
h9_sample(h = NULL, samplesize = NULL, withReplacement = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{samplesize}{The size of the sample as a percentage of the size of the input dataframe}

\item{withReplacement}{Allow or disallow sampling of the same row more than once. Default- True}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Trim a dataset down to a random sample of its rows
}
