% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_sankey}
\alias{h9_sankey}
\title{Sankey}
\usage{
h9_sankey(
  h,
  source = NULL,
  target = NULL,
  value = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{source}{}

\item{target}{}

\item{value}{}

\item{palette}{}

\item{fontsize}{}

\item{marginleft}{}

\item{marginbottom}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Show the flow with Captain Sankey's signature diagram
}
