% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_scatter}
\alias{h9_scatter}
\title{Scatter}
\usage{
h9_scatter(
  h,
  x = NULL,
  y = NULL,
  color = NULL,
  size = NULL,
  label = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{color}{}

\item{size}{}

\item{label}{}

\item{palette}{}

\item{fontsize}{}

\item{marginleft}{}

\item{marginbottom}{}
}
\value{
A list with the pipeline specification.
}
\description{
A tried and true classic
}
