% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_scatter_chart}
\alias{h9_scatter_chart}
\title{Scatter}
\usage{
h9_scatter_chart(
  h,
  x = NULL,
  y = NULL,
  color = NULL,
  size = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The column containing the x coordinates of the marks}

\item{y}{The column containing the y coordinates of the marks}

\item{color}{The column that should be used to group the marks into different colors}

\item{size}{The column the marks should be propotional in area to.}

\item{palette}{Colors to use for the different levels of the y variable. Should be one of the valid d3.js color palettes.}

\item{fontsize}{The size of the font to be used}

\item{marginleft}{The left margin}

\item{marginbottom}{The Bottom margin}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
A tried and true classic
}
