% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_sheet}
\alias{h9_sheet}
\title{Sheet}
\usage{
h9_sheet(uid, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a spreadsheet to collect data
}
