% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_core.R
\name{h9_show}
\alias{h9_show}
\title{Hal9 Show}
\usage{
h9_show(
  h,
  width = NULL,
  height = NULL,
  environment = "devel",
  iframe = TRUE,
  version = "0.3.5",
  mode = c("run", "design"),
  ...
)
}
\arguments{
\item{h}{A h9 object}

\item{width}{Width of the widget container.}

\item{height}{Height of the widget container.}

\item{environment}{The environment to use, production by default. Use 'devel' for
development and 'local' for local developer environment.}

\item{iframe}{Should the application run inside an iframe to protect styles and
resources from the parent? Defaults to 'TRUE'.}

\item{version}{The version of the runtime to use.}

\item{mode}{Should the application be run or designed? Valid values "run" or
"design", defaults to "run" unless the application is empty.}

\item{...}{Additional parameters}
}
\description{
Shows a h9 object.
}
