% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_slice}
\alias{h9_slice}
\title{Slice}
\usage{
h9_slice(h, start = NULL, end = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{start}{The starting index(included)}

\item{end}{The ending index(not included)}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Keep only a contiguous subset of rows
}
