% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{h9_slider}
\alias{h9_slider}
\title{Slider}
\usage{
h9_slider(h, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{...}{Other h9 parameters.}

\item{min}{}

\item{max}{}

\item{step}{}
}
\value{
A list with the pipeline specification.
}
\description{
Embed an slider element
}
