% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{h9_sma}
\alias{h9_sma}
\title{SMA}
\usage{
h9_sma(h, source = NULL, window = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{source}{Additional step parameter.}

\item{window}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Calculate the moving average
}
