% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_sort}
\alias{h9_sort}
\title{Sort}
\usage{
h9_sort(h, field = NULL, order = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{field}{The list of columns to sort by}

\item{order}{The order in which to sort, default - ascending}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Sort rows of data based on the values in selected columns
}
