% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h9_stop}
\alias{h9_stop}
\title{Stop}
\usage{
h9_stop(h, expression = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{expression}{A JavaScript expression, which may use the 'outputs' dictionary which references outputs produced by previous steps by name.}
}
\value{
A list with the pipeline specification.
}
\description{
Stops pipeline execution conditionally
}
