% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transforms.R
\name{h9_summarize}
\alias{h9_summarize}
\title{Summarize}
\usage{
h9_summarize(h = NULL, group = NULL, field = NULL, summarizer = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{group}{The list of columns by which to group}

\item{field}{The list of columns who's values to collect}

\item{summarizer}{The summarizer method to aggregate the values collected, default- count}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Select columns to aggregate and others to keep unmodified
}
