% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_textarea}
\alias{h9_textarea}
\title{Multi-line Text Input}
\usage{
h9_textarea(uid, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a multi-line text input control
}
