% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_treemap}
\alias{h9_treemap}
\title{Treemap}
\usage{
h9_treemap(
  h,
  label = NULL,
  size = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{label}{}

\item{size}{}

\item{palette}{}

\item{fontsize}{}

\item{marginleft}{}

\item{marginbottom}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Choose a column to map into colored blocks, and another to further subdivide by size
}
