% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h9_upscaler}
\alias{h9_upscaler}
\title{Upscale}
\usage{
h9_upscaler(h, originalImg = NULL, model = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{originalImg}{}

\item{model}{}
}
\value{
A list with the pipeline specification.
}
\description{
Magnify an image with AI enhancement
}
