% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_waffle}
\alias{h9_waffle}
\title{Waffle}
\usage{
h9_waffle(h, x = NULL, y = NULL, wafflesizelabel = NULL, palette = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{wafflesizelabel}{}

\item{palette}{}
}
\value{
A list with the pipeline specification.
}
\description{
A square pie chart; still not edible
}
