% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_waffle_chart}
\alias{h9_waffle_chart}
\title{Waffle}
\usage{
h9_waffle_chart(
  h,
  x = NULL,
  y = NULL,
  wafflesizelabel = NULL,
  palette = NULL,
  rebind = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{Additional step parameter.}

\item{y}{Additional step parameter.}

\item{wafflesizelabel}{Additional step parameter.}

\item{palette}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
A square pie chart; still not edible
}
