% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_waterfall}
\alias{h9_waterfall}
\title{Waterfall}
\usage{
h9_waterfall(
  h,
  x = NULL,
  y = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{fontsize}{}

\item{marginleft}{}

\item{marginbottom}{}
}
\value{
A list with the pipeline specification.
}
\description{
Lay out cascades of data next to a towering waterfall representing the total
}
