% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_webscrape_images}
\alias{h9_webscrape_images}
\title{Web Images}
\usage{
h9_webscrape_images(h, url = NULL, minSize = NULL, scrollIters = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{Additional step parameter.}

\item{minSize}{Additional step parameter.}

\item{scrollIters}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Scrape images from a website by using a Hal9 server worker
}
