% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_webscrape_selectors}
\alias{h9_webscrape_selectors}
\title{Web Selectors}
\usage{
h9_webscrape_selectors(
  h,
  url = NULL,
  className = NULL,
  columnName = NULL,
  hasHeader = NULL,
  scrollIters = NULL,
  scrollClick = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{}

\item{className}{}

\item{columnName}{}

\item{hasHeader}{}

\item{scrollIters}{}

\item{scrollClick}{}
}
\value{
A list with the pipeline specification.
}
\description{
Scrape data matching CSS or DOM selectors on a website by using a Hal9 server worker
}
