% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_webscrape_table}
\alias{h9_webscrape_table}
\title{Web Table}
\usage{
h9_webscrape_table(
  h,
  url = NULL,
  text = NULL,
  hasHeader = NULL,
  scrollIters = NULL,
  scrollClick = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{Additional step parameter.}

\item{text}{Additional step parameter.}

\item{hasHeader}{Additional step parameter.}

\item{scrollIters}{Additional step parameter.}

\item{scrollClick}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Scrape a table from a website by using a Hal9 server worker
}
