% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_website}
\alias{h9_website}
\title{Website}
\usage{
h9_website(uid, site = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{...}{Other h9 parameters.}
}
\description{
Embed a website and load a URL
}
