% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_wordcloud}
\alias{h9_wordcloud}
\title{Wordcloud}
\usage{
h9_wordcloud(h, label = NULL, size = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{label}{}

\item{size}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Make the most frequent words in a dataset stand out in a trendy way
}
