% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_visualizations.R
\name{h9_wordcloud_chart}
\alias{h9_wordcloud_chart}
\title{Wordcloud}
\usage{
h9_wordcloud_chart(h = NULL, label = NULL, size = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{label}{The column containing words to be added to the cloud}

\item{size}{The column the size of the words should be in propotion to}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Make the most frequent words in a dataset stand out in a trendy way
}
