% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h9_zip_to_map}
\alias{h9_zip_to_map}
\title{Coordinates from Zip Code}
\usage{
h9_zip_to_map(h, zipcode = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{zipcode}{}
}
\value{
A list with the pipeline specification.
}
\description{
Create latitude and longitude columns from a US zip code column
}
