% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal9_publish.R
\name{hal9_publish}
\alias{hal9_publish}
\title{Publish hal9 app in RPubs}
\usage{
hal9_publish(hal9_expr, title = "My hal9 app", desc = NULL, id = NULL)
}
\arguments{
\item{hal9_expr}{R A hal9 htmlWidget}

\item{id}{If this upload is an update of an existing document then the
id parameter should specify the document id to update. Note that the id
is provided as an element of the list returned by successful calls to
rpubsUpload.}
}
\description{
Publish hal9 app in RPubs
}
