% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal9_render.R
\name{hal9_render}
\alias{hal9_render}
\title{Render hal9 app to HTML file using RMarkdown}
\usage{
hal9_render(hal9_expr, file = "hal9_app.html")
}
\arguments{
\item{hal9_expr}{R A hal9 htmlWidget}

\item{file}{(optional) Name of output file}
}
\value{

}
\description{
Render hal9 app to HTML file using RMarkdown
}
