% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal9_to_html.R
\name{hal9_to_html}
\alias{hal9_to_html}
\title{Render hal9 app to HTML file using RMarkdown}
\usage{
hal9_to_html(
  hal9_expr,
  file = "hal9_app.html",
  title = "My hal9 app",
  desc = NULL
)
}
\arguments{
\item{hal9_expr}{R A hal9 htmlWidget}

\item{file}{(optional) Name of output file}
}
\description{
Render hal9 app to HTML file using RMarkdown
}
