% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_bubble_chart}
\alias{h9_bubble_chart}
\title{Bubbles}
\usage{
h9_bubble_chart(h, label = NULL, size = NULL, palette = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{label}{}

\item{size}{}

\item{palette}{}
}
\value{
A list with the pipeline specification.
}
\description{
Bigger values -> bigger bubbles
}
