% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal9.R
\name{h9_create}
\alias{h9_create}
\title{Hal9 Create}
\usage{
h9_create(
  width = "100\%",
  height = "100\%",
  environment = "cloud",
  iframe = TRUE,
  ...
)
}
\arguments{
\item{width}{Width of the widget container.}

\item{height}{Height of the widget container.}

\item{environment}{The environment to use, production by default. Use 'devel' for
development and 'local' for local developer environment.}

\item{iframe}{Should the application run inside an iframe to protect styles and
resources from the parent? Defaults to 'TRUE'.}

\item{...}{Additional parameters}
}
\description{
Create a h9 object.
}
