% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transforms.R
\name{h9_drop}
\alias{h9_drop}
\title{Drop}
\usage{
h9_drop(h = NULL, columns = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{columns}{The list of columns to remove}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Remove specific columns
}
