% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_drop}
\alias{h9_drop}
\title{Drop}
\usage{
h9_drop(h, columns = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{columns}{The list of columns to remove}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Remove specific columns
}
