% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{h9_dropdown}
\alias{h9_dropdown}
\title{Dropdown}
\usage{
h9_dropdown(h, values = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{values}{}
}
\value{
A list with the pipeline specification.
}
\description{
Embed an HTML <select> element
}
