% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_dropdown}
\alias{h9_dropdown}
\title{Dropdown}
\usage{
h9_dropdown(h = NULL, values = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{values}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed an HTML <select> element
}
