% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_dropdown}
\alias{h9_dropdown}
\title{Dropdown}
\usage{
h9_dropdown(uid, values = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{values}{Additional step parameter.}

\item{...}{Other h9 parameters.}

\item{h}{An optional h9 object created by h9_create.}
}
\description{
Embed a dropdown control
}
