% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{h9_file}
\alias{h9_file}
\title{File Input}
\usage{
h9_file(h, dataType = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{dataType}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed a file input control
}
