% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_file}
\alias{h9_file}
\title{File Input}
\usage{
h9_file(uid, dataType = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{...}{Other h9 parameters.}
}
\description{
Embed a file input control
}
