% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{h9_fold}
\alias{h9_fold}
\title{Fold}
\usage{
h9_fold(h, gather = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{gather}{the list of columns to convert into key-value pairs}
}
\value{
A list with the pipeline specification.
}
\description{
Fold one or more columns into a pair of key-value columns
}
