% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{h9_import_reddit}
\alias{h9_import_reddit}
\title{Reddit}
\usage{
h9_import_reddit(
  h,
  type = NULL,
  sub = NULL,
  query = NULL,
  before = NULL,
  after = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{type}{}

\item{sub}{}

\item{query}{}

\item{before}{}

\item{after}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Search Reddit for submissions or comments
}
