% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{h9_import_sheet}
\alias{h9_import_sheet}
\title{Google Sheets}
\usage{
h9_import_sheet(
  h,
  url = NULL,
  documentAccess = NULL,
  sheet = NULL,
  clientId = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{url}{Additional step parameter.}

\item{documentAccess}{Additional step parameter.}

\item{sheet}{Additional step parameter.}

\item{clientId}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Import data from a Google Sheets spreadsheet
}
