% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_import_stocks}
\alias{h9_import_stocks}
\title{Stocks}
\usage{
h9_import_stocks(
  h,
  stock = NULL,
  statistic = NULL,
  limit = NULL,
  apiKey = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{stock}{Additional step parameter.}

\item{statistic}{Additional step parameter.}

\item{limit}{Additional step parameter.}

\item{apiKey}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Get stock market data provided by Financial Modeling Prep
}
