% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_line_chart}
\alias{h9_line_chart}
\title{Line}
\usage{
h9_line_chart(
  h,
  x = NULL,
  y = NULL,
  palette = NULL,
  domainx = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  rebind = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The column in the dataframe which defines the x coordinate of each vertex of the line}

\item{y}{The column in the dataframe which defines the y coordinate of each vertex of the line}

\item{palette}{Colors to use for the different levels of the y variable. Should be one of the valid d3.js color palettes.}

\item{domainx}{Additional step parameter.}

\item{fontsize}{Additional step parameter.}

\item{marginleft}{Additional step parameter.}

\item{marginbottom}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
For use with sorted X axis values only...unless you like scribbles
}
