% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_import.R
\name{h9_load}
\alias{h9_load}
\title{DataFrame}
\usage{
h9_load(h = NULL, dataset = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{dataset}{The dataframe to load}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Loads a dataframe
}
