% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_map_chart}
\alias{h9_map_chart}
\title{Map}
\usage{
h9_map_chart(h, lon = NULL, lat = NULL, size = NULL, label = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{lon}{The column that contains the values that should be interpreted as the Longitude of vertex}

\item{lat}{The column that contains the values that should be interpreted as the Longitude of vertex}

\item{size}{The column propotional to which the points size should be.}

\item{label}{The column which contains the labels of each of the verticies.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Clearly map out Earthly datasets that have columns for latitude and longitude
}
