% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_number}
\alias{h9_number}
\title{Number Input}
\usage{
h9_number(h = NULL, label = NULL, ...)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{label}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed a number input control
}
