% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_number}
\alias{h9_number}
\title{Number}
\usage{
h9_number(uid, label = NULL, value = NULL, on_update = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element}

\item{label}{Specifies the label for this numberinput}

\item{value}{Specifies the value for this numberinput}

\item{on_update}{Function to call when the value is changed updates}

\item{...}{Other h9 parameters.}
}
\description{
Embed a number input control
}
