% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_number}
\alias{h9_number}
\title{Number Input}
\usage{
h9_number(uid, paramName = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{paramName}{Additional step parameter.}

\item{...}{Other h9 parameters.}

\item{h}{An optional h9 object created by h9_create.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed a number input control
}
