% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_plotly}
\alias{h9_plotly}
\title{Plotly}
\usage{
h9_plotly(
  h,
  x = NULL,
  y = NULL,
  chartType = NULL,
  dataSizes = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{The column that determines the x coordinates in the cartesian plane of the marks}

\item{y}{The column that determines the y coordinates in the cartesian plane of the marks}

\item{chartType}{The chart to be constructed. Currently accepts one of lines, scatter, barChart, fillArea, histogram, twoHistogram}

\item{dataSizes}{The size of the marks}

\item{palette}{Colors to use for the different levels of the y variable. Should be one of the valid d3.js color palettes.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Choose from a variety of Plotly visualizations
}
