% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_radial_bars}
\alias{h9_radial_bars}
\title{Radial}
\usage{
h9_radial_bars(h, x = NULL, y = NULL, palette = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{palette}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
A circular bar chart
}
