% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h9_read_json}
\alias{h9_read_json}
\title{JSON}
\usage{
h9_read_json(h, file = NULL, extract = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{file}{}

\item{extract}{}
}
\value{
A list with the pipeline specification.
}
\description{
Import a JSON formatted dataset
}
