% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{h9_regression_chart}
\alias{h9_regression_chart}
\title{Regression}
\usage{
h9_regression_chart(
  h,
  x = NULL,
  y = NULL,
  type = NULL,
  predictions = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{x}{}

\item{y}{}

\item{type}{}

\item{predictions}{}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Fit a regression model to a dataset to predict future values
}
