% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_visualizations.R
\name{h9_regression_chart}
\alias{h9_regression_chart}
\title{Regression}
\usage{
h9_regression_chart(
  h = NULL,
  x = NULL,
  y = NULL,
  type = NULL,
  predictions = NULL,
  ...
)
}
\arguments{
\item{h}{An optional h9 object created by h9_create.}

\item{x}{Additional step parameter.}

\item{y}{Additional step parameter.}

\item{type}{Additional step parameter.}

\item{predictions}{Additional step parameter.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Fit a regression model to a dataset to predict future values
}
