% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h9_line_chart.R, R/h9_scatter_chart.R
\name{h9_scatter_chart}
\alias{h9_scatter_chart}
\title{Add hal9 line chart step}
\usage{
h9_scatter_chart(h, ...)

h9_scatter_chart(h, ...)
}
\arguments{
\item{h}{A hal9 pipeline}

\item{...}{Unsupported. Used to pass direct parameters do hal9_add_step}

\item{x}{(optional) Character. Variable name for x axis}

\item{y}{(optional) Character. Variable name for y axis}

\item{color}{(optional) Character. Variable name for color}

\item{pallette}{(optional) Character. Color pallete if color is used}
}
\value{



}
\description{
Add hal9 line chart step

Add hal9 scatter chart step
}
