% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_slider}
\alias{h9_slider}
\title{Slider}
\usage{
h9_slider(uid, value = NULL, min = NULL, max = NULL, step = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{...}{Other h9 parameters.}
}
\description{
Embed an slider element
}
