% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{h9_slider}
\alias{h9_slider}
\title{Slider}
\usage{
h9_slider(
  h,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  rebind = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{value}{Additional step parameter.}

\item{min}{Additional step parameter.}

\item{max}{Additional step parameter.}

\item{step}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed an slider element
}
