% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_slider}
\alias{h9_slider}
\title{Slider}
\usage{
h9_slider(uid, value = NULL, min = NULL, max = NULL, step = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{value}{Additional step parameter.}

\item{min}{Additional step parameter.}

\item{max}{Additional step parameter.}

\item{step}{Additional step parameter.}

\item{...}{Other h9 parameters.}

\item{h}{An optional h9 object created by h9_create.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed an slider element
}
