% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{h9_textbox}
\alias{h9_textbox}
\title{Text Input}
\usage{
h9_textbox(h, paramName = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{paramName}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed a text input control
}
