% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_controls.R
\name{h9_textbox}
\alias{h9_textbox}
\title{Text Input}
\usage{
h9_textbox(uid, label = NULL, value = NULL, ...)
}
\arguments{
\item{uid}{Unique identifier for this element.}

\item{label}{Additional step parameter.}

\item{value}{Additional step parameter.}

\item{...}{Other h9 parameters.}

\item{h}{An optional h9 object created by h9_create.}
}
\description{
Embed a text input control
}
