% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{h9_treemap}
\alias{h9_treemap}
\title{Treemap}
\usage{
h9_treemap(
  h,
  label = NULL,
  size = NULL,
  palette = NULL,
  fontsize = NULL,
  marginleft = NULL,
  marginbottom = NULL,
  ...
)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{label}{The column to be used as labels and colorcode the rectangular sectors}

\item{size}{The column to which the area of the rectangular sections should be in propotion to}

\item{palette}{the D3 Palette to determine the color scheme to use}

\item{fontsize}{The size of the font}

\item{marginleft}{The left margin}

\item{marginbottom}{The bottom margin}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Choose a column to map into colored blocks, and another to further subdivide by size
}
