% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{h9_website}
\alias{h9_website}
\title{Website}
\usage{
h9_website(h, site = NULL, rebind = NULL, ...)
}
\arguments{
\item{h}{A h9 object created by h9_create.}

\item{site}{Additional step parameter.}

\item{rebind}{A list of rebindings to apply to this function parameters.}

\item{...}{Other h9 parameters.}
}
\value{
A list with the pipeline specification.
}
\description{
Embed a website and load a URL
}
